MAD initial file splits
