+++ src/st/mad/e_misc.c
